refactor: create separate 'contexts' folder
