+++ src/contexts/account/useAddress.ts
@@ -1,6 +1,6 @@
 import { useCelo } from '@celo/react-celo';
 
-export const useAddress = () => {
+export const useAccountAddress = () => {
   const { address: _address } = useCelo();
   const address = _address ?? null;
   const isConnected = !!address;
