+++ src/features/stake/Stake.tsx
@@ -1,8 +1,8 @@
+import { useAccountContext } from 'src/contexts/account/AccountContext';
 import { useStaking } from 'src/features/stake/useStaking';
 import { SwapForm } from 'src/features/swap/SwapForm';
 import { Switcher } from 'src/features/swap/Switcher';
 import { CenteredLayout } from 'src/layout/CenteredLayout';
-import { useAccountContext } from 'src/providers/AccountProvider';
 import toast from 'src/services/toast';
 import { Celo, toCeloWei } from 'src/utils/tokens';
