+++ src/features/swap/SwapForm.tsx
@@ -4,10 +4,10 @@ import { FormEventHandler, useCallback, useState } from 'react';
 import NumberFormat, { NumberFormatValues } from 'react-number-format';
 import { ThemedIcon } from 'src/components/icons/ThemedIcon';
 import { INPUT_DECIMALS } from 'src/config/consts';
+import { useExchangeContext } from 'src/contexts/exchange/ExchangeContext';
 import { CostsSummary } from 'src/features/swap/CostsSummary';
 import { ReceiveSummary } from 'src/features/swap/ReceiveSummary';
 import { TokenCard } from 'src/features/swap/TokenCard';
-import { useExchangeContext } from 'src/providers/ExchangeProvider';
 import { CeloWei, fromWeiRounded, StCeloWei, Token } from 'src/utils/tokens';
 import { BalanceTools } from './BalanceTools';
 import { SubmitButton } from './SubmitButton';
